% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_manipulation.R
\name{break_filename}
\alias{break_filename}
\title{Divide a filename into multiple variables that have been assigned to the vessel}
\usage{
break_filename(filename)
}
\arguments{
\item{filename}{the filename to be broken up into sections}
}
\value{
a table containing the content of the variable that has been broken up
}
\description{
Divide a filename into multiple variables that have been assigned to the vessel
}
\examples{

# broken = break_filename(filename)

# make_filename(list = broken)
}
