% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_physiological_paramaters.R
\name{calculate_physiological}
\alias{calculate_physiological}
\title{Title}
\usage{
calculate_physiological(cont, pixel_mm = 73, FPS = 22.8, mean = TRUE)
}
\arguments{
\item{mean}{}
}
\value{

}
\description{
Title
}
