% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_manipulation.R
\name{collect_filename}
\alias{collect_filename}
\title{Collect the variables needed to make a file name from the user using a popup}
\usage{
collect_filename(animal = "NS", treatment = "NS", video = "NS", roi = "NS")
}
\arguments{
\item{animal}{the animal that the name should be generated for}

\item{treatment}{type of treatment the animal received}

\item{video}{id of the video which was captured}

\item{roi}{the region of interest selected}
}
\value{
a list containing the critical filename values collected from the user
}
\description{
Collect the variables needed to make a file name from the user using a popup
}
\examples{
# Run in interactive mode only
# make_filename(list = collect_filename())
}
