% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_documentation.r
\docType{data}
\name{example_vessel}
\alias{example_vessel}
\title{Example lymphatic width dataset}
\format{
A data frame with 245,230 rows and 5 variables:
\describe{
  \item{X.1}{identification number of each row}
  \item{y}{y position in the image}
  \item{p_width}{width of the vessel at that position, in pixels}
  \item{excluded}{was that row excluded due to an air bubble}
  \item{filename}{which frame was the pixel row acquired from}
  ...
}
}
\source{
Collected for this package by Peter Russell (2021)
}
\usage{
example_vessel
}
\description{
A data set containing the widths of a test vessel in each frame of a video. Identical
in format to that produced by select_roi and threshold_vessel
}
\keyword{datasets}
