% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_peaks_v3.R
\name{find_contraction_events}
\alias{find_contraction_events}
\title{Find peaks in a vascular time series}
\usage{
find_contraction_events(
  input_vector,
  kband = 30,
  nups = 10,
  min_change = 0.25,
  min_dist = 10,
  plot = FALSE,
  pixel_scale = 73,
  time_scale = 22.5
)
}
\arguments{
\item{input_vector}{vector of values to analyze}

\item{kband}{K smoothing window to apply to the data}

\item{nups}{number of increases before and after the dataset to threshold on}

\item{min_change}{minimum size of change to be termed significant}

\item{min_dist}{Minimum distance between the minima}

\item{plot}{should a plot be returned along with the detected peaks}
}
\value{
A vector of peaks detected
}
\description{
Find peaks in a vascular time series
}
\examples{
i = 1
# Test to come

}
