% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import_bulk.R
\name{import_all_folder}
\alias{import_all_folder}
\title{Parallel find all data code}
\usage{
import_all_folder(csv_files)
}
\arguments{
\item{csv_files}{}
}
\value{

}
\description{
Parallel find all data code
}
