% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import_bulk.R
\name{import_file}
\alias{import_file}
\title{Import a saved file}
\usage{
import_file(filename)
}
\arguments{
\item{filename}{the csv file to import from}
}
\value{
a cleaned up data frame
}
\description{
Import a saved file
}
\examples{
i = 1


}
