% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import_bulk.R
\name{import_folder_bin}
\alias{import_folder_bin}
\title{Parallel mass csv import code}
\usage{
import_folder_bin(current_dir, y_bin = 30)
}
\arguments{
\item{csv_files}{list of csv files to import}
}
\value{
A bulk list of imported csv files
}
\description{
Parallel mass csv import code
}
\examples{

# Test folder to come


import_folder_bin("P:\\\\Full Dataset 2\\\\\\\\\SH28S1")

current_dir = "P:\\\\Full Dataset 2\\\\SH28S1"

}
