% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filename_manipulation.R
\name{make_filename}
\alias{make_filename}
\title{Make a file name that identifies an animal}
\usage{
make_filename(
  animal = "NS",
  treatment = "NS",
  video = "NS",
  roi = "NS",
  list = NULL,
  extension = ""
)
}
\arguments{
\item{animal}{animal id}

\item{treatment}{treatment type}

\item{video}{source video which was measured}

\item{roi}{region of interest imaged}

\item{list}{a list containing all the data, for convenience when being called from other functions}

\item{extension}{the filename extension required}
}
\value{
the name of a file to be used
}
\description{
Make a file name that identifies an animal
}
\examples{
# filename = make_filename("1", "Test", extension = ".csv")

}
