% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{make_matrix}
\alias{make_matrix}
\title{Tile multiple images into a single image}
\usage{
make_matrix(output_list, width = 2)
}
\arguments{
\item{output_list}{list of images to turn into a matrix}

\item{width}{matrix width in images, default 2}
}
\value{
an image, with each input arranged in a matrix
}
\description{
Tile multiple images into a single image
}
\examples{
imagelist = imager::imlist(imager::boats, imager::boats, imager::boats)
matrix = make_matrix(imagelist)
plot(matrix)

}
