% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{output_dir}
\alias{output_dir}
\title{Set the output directory}
\usage{
output_dir(set = NULL, use_default = FALSE, set_default = FALSE)
}
\arguments{
\item{set}{The directory to set to}

\item{use_default}{Should the default value be used, or the system value}

\item{set_default}{Should the system value be updated}
}
\value{
The file path to export to
}
\description{
Set the output directory
}
