% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bulk_Graphing_V2.R
\name{plot_heatmap}
\alias{plot_heatmap}
\title{Plot a heatmap showing a vessel's contraction over time}
\usage{
plot_heatmap(heatmap_file)
}
\arguments{
\item{heatmap_file}{a vmeasur file of vessel diameters over time}
}
\value{
a ggplot2 heatmap
}
\description{
Plot a heatmap showing a vessel's contraction over time
}
\examples{

#heatmap_file = file.choose()
#plot_heatmap(heatmap_file)
}
