% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bulk_Graphing_V2.R
\name{plot_line}
\alias{plot_line}
\title{Plot a line graph of a vessel's change in diameter over time}
\usage{
plot_line(widths_file)
}
\arguments{
\item{widths_file}{a vmeasur created file showing the widths of the vessel over time}
}
\value{
a ggplot2 of the vessel
}
\description{
Plot a line graph of a vessel's change in diameter over time
}
\examples{

#widths_file = file.choose()
#plot_heatmap(heatmap_file)
}
