% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bulk_Graphing_V2.R
\name{quanitfy_mean_width_sections}
\alias{quanitfy_mean_width_sections}
\title{Title}
\usage{
quanitfy_mean_width_sections(widths_file)
}
\arguments{
\item{widths_file}{}
}
\value{

}
\description{
Title
}
