% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantify_directory.R
\name{quantify_directory}
\alias{quantify_directory}
\title{Quantify the content of an entire directory of sub-directories at once}
\usage{
quantify_directory(target_folder)
}
\arguments{
\item{target_folder}{The folder to quantify the readings in}
}
\value{
A PDF file for each directory quantified, showing the quantification
}
\description{
Quantify the content of an entire directory of sub-directories at once
}
