% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bulk_Graphing_V2.R
\name{quantify_folder}
\alias{quantify_folder}
\title{Quantify all ROI in a single folder}
\usage{
quantify_folder(folder)
}
\arguments{
\item{folder}{The folser to quantify}
}
\value{

}
\description{
Quantify all ROI in a single folder
}
