% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bulk_Graphing_V2.R
\name{quantify_folders}
\alias{quantify_folders}
\title{Quantify multiple animals at the same time}
\usage{
quantify_folders(working_folder = tk_choose.dir())
}
\arguments{
\item{working_folder}{The folder containing the completed analysis}
}
\value{
saves a pdf of a series of quantified folders in the root directory
}
\description{
Quantify multiple animals at the same time
}
\examples{
# Used interactivley only

}
