% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roi_select.R
\name{select_roi}
\alias{select_roi}
\title{Select a ROI from a video file}
\usage{
select_roi()
}
\value{
Saves an annotated AVI and CSV file in the same directory as the video.
Will also output and copy the parameters used to create the video.
}
\description{
This function provides a graphical tool to walk the user through selecting a ROI from an AVI video.
}
\examples{

\dontrun{
   select_roi()
}

}
