% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import_bulk.R
\name{summarise_import_file_bin}
\alias{summarise_import_file_bin}
\title{Title}
\usage{
summarise_import_file_bin(fulldata_grouped, file_location)
}
\arguments{
\item{fulldata_grouped}{}
}
\value{

}
\description{
Title
}
