% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_image}
\alias{threshold_image}
\title{Threshold an image}
\usage{
threshold_image(file_path, threshold, min_area = 100)
}
\arguments{
\item{file_path}{Path to the image for quantification}

\item{threshold}{Threshold to set for the inverse channel. Range 0-1.}
}
\value{
A list containing 1) the processed threshold image and 2) calculated widths
}
\description{
Threshold an image
}
\examples{
# No examples

}
