% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Scale_detection.R
\name{calibrate_pixel_size}
\alias{calibrate_pixel_size}
\title{Calibrate the pixel size using a test image}
\usage{
calibrate_pixel_size(file_path = NULL)
}
\arguments{
\item{file_path}{The path to the file}
}
\value{
A graphical representation of
}
\description{
Calibrate the pixel size using a test image
}
\examples{
# file = paste(system.file(package = "vmeasur"), "extdata/mm_scale.jpg", sep = "/")
# calibrate_pixel_size(file)

}
