% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_import_bulk.R
\name{import_folder_bin}
\alias{import_folder_bin}
\title{Parallel mass csv import code}
\usage{
import_folder_bin(current_dir, y_bin = 30)
}
\arguments{
\item{current_dir}{list of csv files to import}

\item{y_bin}{number of pixels to put in each ybin}
}
\value{
A bulk list of imported csv files
}
\description{
Parallel mass csv import code
}
\examples{
# Not applicable to CRAN

}
