% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bulk_Graphing_V2.R
\name{quantify_folder}
\alias{quantify_folder}
\title{Quantify all ROI in a single folder}
\usage{
quantify_folder(folder, kband = 40)
}
\arguments{
\item{folder}{The folder to quantify}

\item{kband}{The k value to use in smoothing the data}
}
\value{
Saves a PDF file of the analyasis generated
}
\description{
Quantify all ROI in a single folder
}
\examples{
# Select a folder and it will be quantified


}
