% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholding.R
\name{threshold_apply}
\alias{threshold_apply}
\title{Title}
\usage{
threshold_apply(
  threshold = 0.5,
  roi_name = "test",
 
    video_path = "\\\\\\\\files.auckland.ac.nz\\\\research\\\\ressci202000061-PROM-study\\\\Version 5\\\\image826.avi",
  radians = 0.217604550320612,
  xlength = 60,
  ylength = 243,
  xstart = 696,
  ystart = 323
)
}
\arguments{
\item{ystart}{}
}
\value{

}
\description{
Title
}
